moved check for user answer to one file in command utils
